Moved testing commands to a separate Cog in testing_commands.py
This will keep the commands organized.
