+++ ping_list.txt
@@ -1,4 +1,3 @@
-387
 367
 611
 450
@@ -18,3 +17,4 @@
 457
 493
 471
+296
